Prefer standard integer types
This (huge) commit also fixes an issue with sprite ordering that was
due to the layer struct actually being larger (more bytes) than it was
supposed to be.

Cleaning up the entire project, considering and specifying just how
many bytes each variable require helped in identifying that issue. It
is now much more clear across the board; for example, resource id's do
not need an entire 4 bytes when there's a hard max at how many you can
have (<32!).

* This is also the first push toward shorter naming conventions.

+++ include/SHIZEN/zloader.h
@@ -21,7 +21,7 @@
  *
  * @return A resource id if the resource was loaded successfully, `0` otherwise
  */
-unsigned int
+u8
 z_load(char const * filename);
 
 /**
@@ -30,7 +30,7 @@ z_load(char const * filename);
  * @return `true` if the resource was unloaded successfully, `false` otherwise
  */
 bool
-z_unload(unsigned int resource_id);
+z_unload(u8 resource_id);
 
 SHIZSprite
 z_load_sprite(char const * filename);
@@ -40,10 +40,10 @@ z_load_sprite_src(char const * filename,
                   SHIZRect source);
 
 SHIZSprite
-z_load_sprite_from(unsigned int resource_id);
+z_load_sprite_from(u8 resource_id);
 
 SHIZSprite
-z_load_sprite_from_src(unsigned int resource_id,
+z_load_sprite_from_src(u8 resource_id,
                        SHIZRect source);
 
 SHIZSpriteSheet
@@ -70,8 +70,8 @@ z_load_sprite_from_index(SHIZSpriteSheet spritesheet,
 
 SHIZSprite
 z_load_sprite_from_cell(SHIZSpriteSheet spritesheet,
-                        unsigned int column,
-                        unsigned int row);
+                        u16 column,
+                        u16 row);
 
 SHIZSpriteFont
 z_load_spritefont(char const * filename,
